requirements.txt
Django~=2.0.6
djangorestframework==3.9.0